Change DevInquire return type to c_long
Return from DevInquire is updated to conform with ALP4's technical
document.

+++ src/ALP4.py
@@ -710,7 +710,7 @@ class ALP4(object):
         RETURNS
         -------
 
-        value : c_double
+        value : c_long
                 Value of the requested parameter.
 
         SEE ALSO
@@ -720,7 +720,7 @@ class ALP4(object):
 
         """
 
-        ret = ct.c_double(0)
+        ret = ct.c_long(0)
 
         self._checkError(self._ALPLib.AlpDevInquire(self.ALP_ID, inquireType, ct.byref(ret)), 'Error sending request.')
         return ret.value
